+++ src/main/java/com/elovirta/dita/markdown/renderer/TopicRenderer.java
@@ -3,7 +3,6 @@
  */
 package com.elovirta.dita.markdown.renderer;
 
-import static com.elovirta.dita.markdown.DitaRenderer.RAW_DITA;
 import static com.elovirta.dita.markdown.renderer.Utils.buildAtts;
 import static javax.xml.XMLConstants.XML_NS_URI;
 import static org.dita.dost.util.Constants.*;
@@ -53,8 +52,6 @@ import org.dita.dost.util.DitaClass;
 import org.xml.sax.Attributes;
 import org.xml.sax.InputSource;
 import org.xml.sax.SAXException;
-import org.xml.sax.XMLReader;
-import org.xml.sax.helpers.XMLFilterImpl;
 
 /**
  * A renderer for a set of node types.
@@ -706,7 +703,7 @@ public class TopicRenderer extends AbstractRenderer {
     h.setResult(new SAXResult(fragmentFilter));
     final HtmlParser parser = new HtmlParser();
     parser.setNamePolicy(XmlViolationPolicy.ALLOW);
-    final MathmlNamespaceFilter filter = new MathmlNamespaceFilter(parser);
+    final NamespaceFilter filter = new NamespaceFilter(parser);
     filter.setContentHandler(h);
     try {
       html.setLocation(node);
@@ -715,49 +712,6 @@ public class TopicRenderer extends AbstractRenderer {
       throw new ParseException("Failed to parse HTML: " + e.getMessage(), e);
     }
     html.setDocumentLocator();
-  }
-
-  private static final class MathmlNamespaceFilter extends XMLFilterImpl {
-
-    public MathmlNamespaceFilter(XMLReader parent) {
-      super(parent);
-    }
-
-    public void startElement(String uri, String localName, String qName, Attributes atts) throws SAXException {
-      if (qName.startsWith("m:")) {
-        final String tag = qName.substring(2);
-        if (qName.equals("m:math")) {
-          super.startPrefixMapping("m", "http://www.w3.org/1998/Math/MathML");
-        }
-        super.startElement(uri, tag, qName, atts);
-      } else if (qName.startsWith("svg:")) {
-        final String tag = qName.substring(4);
-        if (qName.equals("svg:svg")) {
-          super.startPrefixMapping("svg", "http://www.w3.org/2000/svg");
-        }
-        super.startElement(uri, tag, qName, atts);
-      } else {
-        super.startElement(uri, localName, qName, atts);
-      }
-    }
-
-    public void endElement(String uri, String localName, String qName) throws SAXException {
-      if (qName.startsWith("m:")) {
-        final String tag = qName.substring(2);
-        super.endElement(uri, tag, qName);
-        if (qName.equals("m:math")) {
-          super.endPrefixMapping("m");
-        }
-      } else if (qName.startsWith("svg:")) {
-        final String tag = qName.substring(2);
-        super.endElement(uri, tag, qName);
-        if (qName.equals("svg:svg")) {
-          super.endPrefixMapping("svg");
-        }
-      } else {
-        super.endElement(uri, localName, qName);
-      }
-    }
   }
 
   private static Stream<Entry<String, Entry<DitaClass, Attributes>>> createHtmlToDita(Entry<String, DitaClass> e) {
